+++ instrumentation/opentelemetry-instrumentation-psycopg/tests/test_psycopg_integration.py
@@ -22,6 +22,9 @@ from psycopg.sql import SQL, Composed
 import opentelemetry.instrumentation.psycopg
 from opentelemetry.instrumentation.psycopg import PsycopgInstrumentor
 from opentelemetry.sdk import resources
+from opentelemetry.semconv._incubating.attributes.db_attributes import (
+    DB_OPERATION_NAME,
+)
 from opentelemetry.test.test_base import TestBase
 
 
@@ -103,18 +106,18 @@ class MockConnection:
 
 
 class MockAsyncConnection(psycopg.AsyncConnection):
-    commit = mock.MagicMock(spec=types.MethodType)
-    commit.__name__ = "commit"
-
-    rollback = mock.MagicMock(spec=types.MethodType)
-    rollback.__name__ = "rollback"
-
     def __init__(self, *args, **kwargs):
         self.cursor_factory = kwargs.pop("cursor_factory", None)
 
     @staticmethod
     async def connect(*args, **kwargs):
         return MockAsyncConnection(**kwargs)
+
+    async def commit(self):
+        pass
+
+    async def rollback(self):
+        pass
 
     def cursor(self, *args, **kwargs):
         if self.cursor_factory:
@@ -441,6 +444,30 @@ class TestPostgresqlIntegration(PostgresqlIntegrationTestMixin, TestBase):
         spans_list = self.memory_exporter.get_finished_spans()
         self.assertEqual(len(spans_list), 1)
 
+    def test_commit(self):
+        PsycopgInstrumentor().instrument(enable_transaction_spans=True)
+
+        cnx = psycopg.connect(database="test")
+        cnx.commit()
+
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 1)
+        span = spans_list[0]
+        self.assertEqual(span.name, "COMMIT")
+        self.assertEqual(span.attributes[DB_OPERATION_NAME], "COMMIT")
+
+    def test_rollback(self):
+        PsycopgInstrumentor().instrument(enable_transaction_spans=True)
+
+        cnx = psycopg.connect(database="test")
+        cnx.rollback()
+
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 1)
+        span = spans_list[0]
+        self.assertEqual(span.name, "ROLLBACK")
+        self.assertEqual(span.attributes[DB_OPERATION_NAME], "ROLLBACK")
+
     @mock.patch("opentelemetry.instrumentation.dbapi.wrap_connect")
     def test_sqlcommenter_enabled(self, event_mocked):
         cnx = psycopg.connect(database="test")
@@ -588,6 +615,34 @@ class TestPostgresqlIntegrationAsync(
             self.assertTrue(mock_span.is_recording.called)
             self.assertFalse(mock_span.set_attribute.called)
             self.assertFalse(mock_span.set_status.called)
+
+        PsycopgInstrumentor().uninstrument()
+
+    async def test_async_commit(self):
+        PsycopgInstrumentor().instrument(enable_transaction_spans=True)
+
+        cnx = await psycopg.AsyncConnection.connect("test")
+        await cnx.commit()
+
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 1)
+        span = spans_list[0]
+        self.assertEqual(span.name, "COMMIT")
+        self.assertEqual(span.attributes[DB_OPERATION_NAME], "COMMIT")
+
+        PsycopgInstrumentor().uninstrument()
+
+    async def test_async_rollback(self):
+        PsycopgInstrumentor().instrument(enable_transaction_spans=True)
+
+        cnx = await psycopg.AsyncConnection.connect("test")
+        await cnx.rollback()
+
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 1)
+        span = spans_list[0]
+        self.assertEqual(span.name, "ROLLBACK")
+        self.assertEqual(span.attributes[DB_OPERATION_NAME], "ROLLBACK")
 
         PsycopgInstrumentor().uninstrument()