+++ instrumentation/opentelemetry-instrumentation-pymysql/tests/test_pymysql_integration.py
@@ -21,7 +21,9 @@ from opentelemetry import trace as trace_api
 from opentelemetry.instrumentation.pymysql import PyMySQLInstrumentor
 from opentelemetry.sdk import resources
 from opentelemetry.semconv._incubating.attributes.db_attributes import (
+    DB_OPERATION_NAME,
     DB_STATEMENT,
+    DB_SYSTEM,
 )
 from opentelemetry.test.test_base import TestBase
 
@@ -482,3 +484,58 @@ class TestPyMysqlIntegration(TestBase):
 
         spans_list = self.memory_exporter.get_finished_spans()
         self.assertEqual(len(spans_list), 1)
+
+    @mock.patch("pymysql.connect")
+    # pylint: disable=unused-argument
+    def test_commit(self, mock_connect):
+        """Test that commit creates a span"""
+        PyMySQLInstrumentor().instrument(enable_transaction_spans=True)
+        cnx = pymysql.connect(database="test")
+        cnx.commit()
+
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 1)
+        span = spans_list[0]
+        self.assertEqual(span.name, "COMMIT")
+        self.assertIs(span.kind, trace_api.SpanKind.CLIENT)
+        self.assertEqual(span.attributes[DB_SYSTEM], "mysql")
+        self.assertEqual(span.attributes[DB_OPERATION_NAME], "COMMIT")
+
+    @mock.patch("pymysql.connect")
+    # pylint: disable=unused-argument
+    def test_rollback(self, mock_connect):
+        """Test that rollback creates a span"""
+        PyMySQLInstrumentor().instrument(enable_transaction_spans=True)
+        cnx = pymysql.connect(database="test")
+        cnx.rollback()
+
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 1)
+        span = spans_list[0]
+        self.assertEqual(span.name, "ROLLBACK")
+        self.assertIs(span.kind, trace_api.SpanKind.CLIENT)
+        self.assertEqual(span.attributes[DB_SYSTEM], "mysql")
+        self.assertEqual(span.attributes[DB_OPERATION_NAME], "ROLLBACK")
+
+    @mock.patch("pymysql.connect")
+    # pylint: disable=unused-argument
+    def test_commit_and_query(self, mock_connect):
+        """Test that both execute and commit create spans"""
+        PyMySQLInstrumentor().instrument(enable_transaction_spans=True)
+        cnx = pymysql.connect(database="test")
+        cursor = cnx.cursor()
+        cursor.execute("SELECT * FROM test")
+        cnx.commit()
+
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 2)
+
+        # First span should be the SELECT
+        select_span = spans_list[0]
+        self.assertEqual(select_span.name, "SELECT")
+        self.assertIs(select_span.kind, trace_api.SpanKind.CLIENT)
+
+        # Second span should be the COMMIT
+        commit_span = spans_list[1]
+        self.assertEqual(commit_span.name, "COMMIT")
+        self.assertIs(commit_span.kind, trace_api.SpanKind.CLIENT)